+++ tests/test_transaction.py
@@ -7,23 +7,29 @@ from packaging.tags import Tag
     "path",
     [
         SNOWBALL_WHEEL,
-        f"/{SNOWBALL_WHEEL}" f"a/{SNOWBALL_WHEEL}",
+        f"/{SNOWBALL_WHEEL}",
+        f"a/{SNOWBALL_WHEEL}",
         f"/a/{SNOWBALL_WHEEL}",
         f"//a/{SNOWBALL_WHEEL}",
     ],
 )
-@pytest.mark.parametrize("protocol", ["https:", "file:", "emfs:", ""])
+@pytest.mark.parametrize(
+    "protocol",
+    ["http:", "https:", "file:", "emfs:", ""],
+)
 def test_parse_wheel_url1(protocol, path):
     pytest.importorskip("packaging")
     from micropip.transaction import WheelInfo
 
     url = protocol + path
     wheel = WheelInfo.from_url(url)
+
+    check_url = url if protocol else "file:///" + path
     assert wheel.name == "snowballstemmer"
     assert str(wheel.version) == "2.0.0"
     assert wheel.sha256 is None
     assert wheel.filename == SNOWBALL_WHEEL
-    assert wheel.url == url
+    assert wheel.url == check_url
     assert wheel.tags == frozenset(
         {Tag("py2", "none", "any"), Tag("py3", "none", "any")}
     )
@@ -48,6 +54,8 @@ def test_parse_wheel_url3():
 
 
 def create_transaction(Transaction):
+    from micropip.package_index import DEFAULT_INDEX_URLS
+
     return Transaction(
         wheels=[],
         locked={},
@@ -59,7 +67,7 @@ def create_transaction(Transaction):
         ctx={},
         ctx_extras=[],
         fetch_kwargs={},
-        index_urls=None,
+        index_urls=DEFAULT_INDEX_URLS,
     )