+++ CHANGELOG.md
@@ -6,10 +6,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `micropip.install` now works with simple http indexes that use relative
+  links when referencing wheels. [#150](https://github.com/pyodide/micropip/pull/150)
+
 ### Added
 
 - Added support for PEP-658.
   [#139](https://github.com/pyodide/micropip/pull/139)
+
+- `micropip.install(index_urls=[...])` parameter now supports the special value
+  `"PYPI"` to refer the `http://pypi.org/simple/` index instead of having to
+  type the full url.
+
+## [0.6.1] - 2024/10/05
+
+### Fixed
+
+- micropip.install and micropip.uninstall now accepts `verbosity=None`
+  which does not overwrite a default log level.
+  [#132](https://github.com/pyodide/micropip/pull/132)
+
+- When multiple index urls are given, micropip.install will now correctly
+  fallback to the next index url when one index url fails to find a package.
+  [#129](https://github.com/pyodide/micropip/pull/129)
 
 ## [0.6.0] - 2024/01/31